+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Heroic Games Launcher application scanner, discovering installed Epic, GOG, Amazon and sideloaded games with box art.
+
 ## [v0.14.5] - 30-07-2026
 
 ### Changed